requirements.txt
streamlit==1.35.0
pandas==2.2.3
psycopg2-binary==2.9.10
plotly==6.2.0
prophet==1.1.6
python-dotenv==1.1.0